Move `ImageDecoder` to `misc/image_decoder.cpp|h`
