+++ src/capture/image/image_saver.cpp
@@ -7,11 +7,11 @@
 #include <cstring>
 #include <vector>
 
-#include "image_decoder.h"
 #include "image_saver.h"
 
 #include "capture/capture.h"
 #include "hardware/video/vga.h"
+#include "misc/image_decoder.h"
 #include "misc/support.h"
 #include "png_writer.h"
 #include "utils/bgrx8888.h"
