+++ src/capture/meson.build
@@ -4,7 +4,6 @@ libcapture_sources = files(
     'capture_midi.cpp',
     'capture_video.cpp',
     'image/image_capturer.cpp',
-    'image/image_decoder.cpp',
     'image/image_saver.cpp',
     'image/image_scaler.cpp',
     'image/png_writer.cpp',
