+++ src/misc/image_decoder.h
